% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LG5data}
\alias{LG5data}
\title{LG5data data}
\format{
The first 10 columns are the variables. The last column is the true 
 classification vector where symbol "0" stands for the contaminating data points.
}
\usage{
data(LG5data)
}
\description{
A data set in dimension 10 with three clusters around affine subspaces
 of common intrinsic dimension. A 10\% background noise is added uniformly 
 distributed in a rectangle containing the three main clusters.
}
\examples{
#--- EXAMPLE 1 ------------------------------------------ 
data (LG5data)
x <- LG5data[, 1:10]
clus <- rlg(x, d = c(2,2,2), alpha=0.1, trace=TRUE)
plot(x, col=clus$cluster+1)
}
\keyword{datasets}
