% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pine}
\alias{pine}
\title{Pinus nigra dataset}
\format{
A data frame containing 362 observations in 2 variables. 
The variables are as follows:

\itemize{
     \item \code{Diameter} Diameter
     \item \code{HT} Height
 }
}
\usage{
data(pine)
}
\description{
To study the growth of the wood mass in a cultivated forest of \emph{Pinus nigra} 
 located in the north of Palencia (Spain), a sample of 362 trees was studied. 
 The data set is made of measurements of heights (in meters), in variable "HT", 
 and diameters (in millimetres), in variable "Diameter", of these trees. 
 The presence of three linear groups can be guessed apart from a small group 
 of trees forming its own cluster with larger heights and diameters one isolated 
 tree with the largest diameter but small height. More details on the 
 interpretation of this dataset in García-Escudero et al (2010).
}
\references{
García-Escudero, L. A., Gordaliza, A., Mayo-Iscar, A., and San Martín, R. (2010). 
 Robust clusterwise linear regression through trimming. 
 \emph{Computational Statistics & Data Analysis}, 54(12), 3057--3069.
}
\keyword{datasets}
