% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tclustICsol.R
\name{plot.tclustICsol}
\alias{plot.tclustICsol}
\title{Plot Method for \code{tclustICsol} Objects}
\usage{
\method{plot}{tclustICsol}(x, whichIC, sol = 1, col, pch, main, sub1, choice, ...)
}
\arguments{
\item{x}{The \code{tclustICsol} object to be displayed}

\item{whichIC}{A character value which specifies which information criteria 
to use. Fpr the possible values for \code{whichIC} see the help of \code{tclustIC}.}

\item{sol}{Which solution to display - a number between 1 and the \code{nsol}
argument of \code{tclustICsol}.}

\item{col}{optional colors to identify the groups. If not specified default 
values will be selected.}

\item{pch}{optional symbols to identify the groups. If not specified default 
values will be selected.}

\item{main}{optional title. The default title shows the information criteria, 
the solution number, number of groups and restriction factor c used and 
indication wheather the solution is true or spurios.}

\item{sub1}{an optional subtitle. The default subtitle shows the type of the 
displaued solution: 'Best in' and 'Stable in'.}

\item{choice}{a numeric vector of length between 2 and the number of 
variables in the input data matrix. If missing, a scatterplot matrix with 
all variables will be shown. If two variables are selected, a simple 
scatterplot of the two selected variables is shown.}

\item{\ldots}{Further (optional) graphical arguments}
}
\description{
Displays one of the solutions, selected by the argument 'sol'.
 The default display is a scatterplot matrix of the data using
 colors and symbols of the observations to identify the groups in the 
 selected solution. If the argument 'choice' is specified and its length 
 is two, a simple scatterplot will be shown.
 The function fails, if \code{store.x = FALSE} is specified in 
 the \code{tclustICsol()} call because the original data matrix is 
 required here.
}
\details{
The plot method for class \code{tclustICsol}.
}
\examples{

 #--- EXAMPLE 1 ------------------------------------------
 \donttest{
 data(geyser2)
 (out <- tclustIC(geyser2, whichIC="MIXMIX", alpha=0.1))

 ##  Show the first two best solutions using as Information criterion MIXMIX
 cat("\nBest solutions using MIXMIX\n")
 outsol <- tclust::tclustICsol(out, whichIC="MIXMIX", nsol=2)
 plot(outsol)
 plot(outsol, choice=c(1,2))
 plot(outsol, choice=c(1,2), xlab="XLAB", ylab="YLAB")
 }

}
