% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eventlog_csv.R
\name{read_eventlog_csv}
\alias{read_eventlog_csv}
\title{Read event log in CSV format}
\usage{
read_eventlog_csv(filename)
}
\arguments{
\item{filename}{path to event log}
}
\value{
tibble having event log
}
\description{
Currently, this function requires case, event and completeTime
columns to be present in event log to be read.
}
