% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_cases.R
\name{summarise_cases}
\alias{summarise_cases}
\title{Summarise cases}
\usage{
summarise_cases(eventlog)
}
\arguments{
\item{eventlog}{event log having case and event columns}
}
\value{
tibble having case, trace_length and unique_activities
columns
}
\description{
Get trace length and number of unique activity
for every case in event log.
}
