% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_eventlog.R
\name{summarise_eventlog}
\alias{summarise_eventlog}
\title{Summarise event log}
\usage{
summarise_eventlog(eventlog)
}
\arguments{
\item{eventlog}{event log}
}
\value{
named vector having summary metrics
}
\description{
Returns summary metrics of event log
}
