\name{swissbank}
\encoding{latin1}
\alias{swissbank}
\title{ SwissBankNotes Data }
\description{ 
Six variables measured on 100 genuine and 100 counterfeit old Swiss 1000-franc bank notes (Flury and Riedwyl, 1988). 
}
\usage{data(swissbank)}
\format{
\describe{
\item{\code{Length}}{ Length of the bank note }
\item{\code{Ht_Left}}{ Height of the bank note, measured on the left }
\item{\code{Ht_Right}}{ Height of the bank note, measured on the right }
\item{\code{IF_Lower}}{ Distance of inner frame to the lower border }
\item{\code{IF_Upper}}{ Distance of inner frame to the upper border }
\item{\code{Diagonal}}{ Length of the diagonal }
}
}
\details{Observations 1-100 are the genuine bank notes and the other 100 observations are the counterfeit bank notes. }
\source{ Flury, B. and Riedwyl, H. (1988). Multivariate Statistics, A Practical Approach, Cambridge University Press. }
