\name{tclust-package}
\encoding{latin1}
\alias{tclust-package}
\docType{package}
\title{ General Trimmed Cluster Analysis }
\description{ A package implementing the cluster algorithm (\code{\link[tclust]{tclust}})as introduced in ..., including some graphical diagnostic tools (\code{\link[tclust]{ctlcurves}} and \code{\link[tclust]{bayesfact}}) }
\details{
\tabular{ll}{
Package: \tab tclust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-13\cr
License: \tab What license is it under?\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Agustin Mayo Iscar,
Heinrich Fritz,
Maintainer: Luis Angel Garcia Escudero <lagarcia@eio.uva.es>
}
%\references{ % How to cite your paper? }
\keyword{ package }
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{ %~~ Optional links to other man pages, e.g. ~~ %~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~ }
%\examples{ %~~ simple examples of the most important functions ~~ }
