% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_cluster_nums.R
\name{top_n_cluster_nums}
\alias{top_n_cluster_nums}
\title{Pick top n clusters}
\usage{
top_n_cluster_nums(cluster_summary, n = 5)
}
\arguments{
\item{cluster_summary}{cluster summary tibble made using
\link{summarise_clusters}}

\item{n}{how many top clusters to pick?}
}
\description{
Simply picks top n rows from cluster summary tibble
}
