% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_distmatrix.R
\name{trace_distmatrix}
\alias{trace_distmatrix}
\title{Compute trace distance matrix}
\usage{
trace_distmatrix(eventlog, method = "lv", ...)
}
\arguments{
\item{method}{string distance method to be used as discussed in
\link{stringdist-metrics}}

\item{...}{further arguments to be passed to \link{seq_distmatrix}}
}
\value{
\code{dist} object having pairwise distances between
traces
}
\description{
This function uses \link{seq_distmatrix} from \code{stringdist} package
to compute pairwise distance matrix.
}
