% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{The wine dataset from the UCI Machine Learning Repository}
\format{
A data frame containing 178 observations in 14 variables (13 numerical and one categorical). 
The variables are as follows:

\itemize{
     \item \code{Alcohol} Alcohol
     \item \code{Malic} Malic acid
     \item \code{Ash} Ash
     \item \code{Alcalinity} Alcalinity of ash 
     \item \code{Magnesium} Magnesium
     \item \code{Phenols} Total phenols
     \item \code{Flavanoids} Flavanoids
     \item \code{Nonflavanoids} Nonflavanoid phenols
     \item \code{Proanthocyanins} Proanthocyanins
     \item \code{Color} Color intensity
     \item \code{Hue} Hue 
     \item \code{Dilution} D280/OD315 of diluted wines.
     \item \code{Proline} Proline
     \item \code{class} The type of wine, into one of three classes, 
         1 (59 obs), 2(71 obs), and 3 (48 obs).
 }
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Wine}
}
\usage{
data(wine)
}
\description{
The wine dataset contains the results of a chemical analysis of wines grown in 
 a specific area of Italy. Three types of wine are represented in the 178 samples, 
 with the results of 13 chemical analyses recorded for each sample. The \code{class} 
 variable is a categoric variable. The data contains no missing values and consits 
 of only numeric data, with a three class target variable (\code{class}) for classification.
}
\examples{
#--- EXAMPLE 1 ------------------------------------------ 
data (wine) 
x <- wine[, -ncol(wine)] 
clus <- tclust(x, k=3, alpha=0.1, nstart=200, niter1=3, niter2=17, 
   nkeep=10, opt="HARD", equal.weights=FALSE, restr.fact=50, trace=TRUE) 
plot (x, col=clus$cluster+1)
 plot(clus)
}
\keyword{datasets}
