% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clusters.R
\name{write_clusters}
\alias{write_clusters}
\title{Write clusters}
\usage{
write_clusters(eventlog, cluster_nums, subdirname = "eventlog_clusters",
  filename_prefix = "cluster_")
}
\arguments{
\item{eventlog}{clustered event log (event log with cluster column)}

\item{cluster_nums}{which cluster numbers to export}

\item{subdirname}{in which sub directory you want to export files?}

\item{filename_prefix}{prefix to attach to each output file}
}
\value{
This is a function just for side effects.
}
\description{
Output CSV files for clusters (one CSV per cluster). Currently this function
only exports case, event and completeTime columns. It also outputs a cluster_summary.csv
file having summary metrics of all clusters.
}
