% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geyser2}
\alias{geyser2}
\title{Old Faithful Geyser Data}
\format{
A data frame containing 272 observations in 2 variables. 
The variables are as follows:

\itemize{
     \item \code{Eruption length} The eruption length in minutes. 
     \item \code{Previous eruption length} The length of the previous eruption in minutes. 
 }
}
\source{
This particular data structure can be obtained by applying the following code
 to the "Old Faithful Geyser" (\code{faithful} data set (Härdle 1991) in the 
 package \code{datasets}):
 \cr
 \code{f1 <- faithful[,1]}\cr
 \code{geyser2 <- cbind(f1[-length(f1)], f1[-1])}\cr
 \code{colnames(geyser2) <- c("Eruption length",}\cr
 \code{                      "Previous eruption length")}
}
\usage{
data(geyser2)
}
\description{
A bivariate data set obtained from the Old Faithful Geyser, containing the 
 eruption length and the length of the previous eruption for 271 eruptions 
 of this geyser in minutes.
}
\references{
García-Escudero, L.A. and Gordaliza, A. (1999). 
 Robustness properties of k-means and trimmed k-means. \emph{Journal of the American Statistical Assoc.}, Vol.94, No.447, 956--969.

 Härdle, W. (1991). \emph{Smoothing Techniques with Implementation in S.}, New York: Springer.
}
\keyword{datasets}
