\name{M5data}
\encoding{latin1}
\alias{M5data}
\title{ Mixture M5 Data }
\description{ 
A bivariate data set obtained from three normal bivariate distributions with different scales and proportions 1:2:2. One of the components is very overlapped with another one. A 10\% background noise is added uniformly distributed in a rectangle containing the three normal components and not very overlapped with the three mixture components. A precise description of the M5 data set can be found in García-Escudero et al. (2008).
}
\usage{data(M5data)}
\format{
    The first two columns are the two variables. The last column is the true classification vector where symbol "0" stands for the contaminating data points.
}
\source{ García-Escudero, L.A.; Gordaliza, A.; Matrán, C. and Mayo-Iscar, A. (2008), "A General Trimming Approach to Robust Cluster Analysis". Annals of Statistics, Vol.36, pp. 1324-1345. Technical report available at http://www.eio.uva.es/inves/grupos/representaciones/trTCLUST.pdf }
