% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swissbank}
\alias{swissbank}
\title{Swiss banknotes data}
\format{
A data frame containing 200 observations in 6 variables. 
The variables are as follows:

\itemize{
     \item \code{Length} Length of the bank note 
     \item \code{Ht_Left} Height of the bank note, measured on the left 
     \item \code{Ht_Right} Height of the bank note, measured on the right 
     \item \code{IF_Lower} Distance of inner frame to the lower border 
     \item \code{IF_Upper} Distance of inner frame to the upper border 
     \item \code{Diagonal} Length of the diagonal 
 }
}
\source{
Flury, B. and Riedwyl, H. (1988). \emph{Multivariate Statistics, A Practical Approach}, Cambridge University Press.
}
\usage{
data(swissbank)
}
\description{
Six variables measured on 100 genuine and 100 counterfeit old Swiss 1000-franc 
 bank notes (Flury and Riedwyl, 1988).
}
\details{
Observations 1--100 are the genuine bank notes and the other 100 observations are the counterfeit bank notes.
}
\keyword{datasets}
