\name{tclust-package}
\encoding{latin1}
\alias{tclust-package}
\docType{package}
\title{ General Trimmed Cluster Analysis }
\description{ A package implementing different (robust) clustering algorithms (\code{\link[tclust]{tclust}}) based on trimming 
and including some graphical diagnostic tools (\code{\link[tclust]{ctlcurves}} and \code{\link[tclust]{DiscrFact}}) }
\details{
\tabular{ll}{
Package: \tab tclust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-13\cr
License: \tab GPL-3\cr}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Agustin Mayo Iscar,
Heinrich Fritz,
Maintainer: Luis Angel Garcia Escudero <lagarcia@eio.uva.es>
}
\references{ 
\enc{García}{Garcia}-Escudero, L.A.; Gordaliza, A.; \enc{Matrán}{Matran}, C. and Mayo-Iscar, A. (2008), "A General Trimming Approach to Robust Cluster Analysis." Annals of Statistics, Vol.36, 1324-1345.\cr
\enc{García}{Garcia}-Escudero, L.A.; Gordaliza, A.; \enc{Matrán}{Matran}, C. and Mayo-Iscar, A. (2010), "A Review of Robust Clustering Methods." Advances in Data Analysis and Classification, Vol.4, 89-109.\cr
Fritz, H.; Garcia-Escudero, L.A.; Mayo-Iscar, A. (2012), "tclust: An R Package for a Trimming Approach to Cluster Analysis". Journal of Statistical Software, 47(12), 1-26. URL http://www.jstatsoft.org/v47/i12/}
\keyword{ package }
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{ %~~ Optional links to other man pages, e.g. ~~ %~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~ }
%\examples{ %~~ simple examples of the most important functions ~~ }
